pyproject.toml
[project]
name = "gs24-crawl-server"
version = "0.1.0"
description = ""
authors = [
    {name = "tjdus",email = "dusml0277@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
    "requests (>=2.32.3,<3.0.0)",
    "konlpy (>=0.6.0,<0.7.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "selenium (>=4.32.0,<5.0.0)",
    "kafka-python (>=2.2.6,<3.0.0)",
    "elasticsearch (>=9.0.1,<10.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "transformers (>=4.51.3,<5.0.0)",
    "torch (>=2.7.0,<3.0.0)",
    "protobuf (>=6.31.0,<7.0.0)",
]
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
